Move EmptyCursor to `core:android:common`
