+++ build-plugin/src/main/kotlin/thunderbird.library.android.gradle.kts
@@ -27,5 +27,6 @@ android {
 
 dependencies {
     implementation(libs.bundles.shared.jvm.main)
+    implementation(libs.bundles.shared.jvm.android)
     testImplementation(libs.bundles.shared.jvm.test)
 }
